api/pyproject.toml
[project]
name = "icd11-api"
version = "0.1.0"
description = "FastAPI backend for ICD-11 exploration"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["ruff"]

[tool.ruff]
line-length = 100
